Fixed back port maven plugin
