+++ rune-maven-plugin/src/main/java/com/regnosys/rosetta/maven/RuneMavenStandaloneBuilderModule.java
@@ -19,10 +19,10 @@ package com.regnosys.rosetta.maven;
 import org.eclipse.xtext.builder.standalone.StandaloneBuilder;
 import org.eclipse.xtext.maven.MavenStandaloneBuilderModule;
 
-public class RosettaMavenStandaloneBuilderModule extends MavenStandaloneBuilderModule {
+public class RuneMavenStandaloneBuilderModule extends MavenStandaloneBuilderModule {
 	@Override
 	protected void configure() {
 		super.configure();
-		bind(StandaloneBuilder.class).to(RosettaStandaloneBuilder.class);
+		bind(StandaloneBuilder.class).to(RuneStandaloneBuilder.class);
 	}
 }
